requirements.txt
fastapi
uvicorn
#playwright
#playwright_stealth
patchright
python-dotenv
fake-useragent
requests
#html2text
markdownify
html_sanitizer